test: split workspace setup helpers
